requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.2.0
pkg-resources==0.0.0
pygame==2.0.1
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
torch==1.8.1

